+++ share/pot/start.sh
@@ -163,11 +163,21 @@ _js_export_ports()
 		if [ "$_pot_port" = "$_port" ]; then
 			_host_port=$( _js_get_free_rnd_port "$_excl_list" )
 		fi
-		_debug "Redirect: from $POT_EXTIF : $_host_port to $_ip : $_port"
+		_debug "Redirect: from $POT_EXTIF : $_host_port to $_ip : $_pot_port"
 		echo "rdr pass on $POT_EXTIF proto tcp from any to $POT_EXTIF port $_host_port -> $_ip port $_pot_port" >> $_pfrules
 		_excl_list="$_excl_list $_host_port"
 	done
 	pfctl -a pot-rdr/$_pname -f $_pfrules
+	_lo_tunnel="$(_get_conf_var "$_pname" "pot.attr.localhost-tunnel")"
+	if [ "$_lo_tunnel" = "YES" ]; then
+		_pdir="${POT_FS_ROOT}/jails/$_pname/"
+		if [ -x "/usr/local/bin/ncat" ]; then
+			cp /usr/local/bin/ncat "$_pdir/ncat-$_pname"
+			daemon -f -p $_pdir/ncat.pid $_pdir/ncat-$_pname -lk "$_host_port" -c "/usr/local/bin/ncat $_ip $_pot_port"
+		else
+			_error "nmap package is missing, localhost-tunnel attribute ignored"
+		fi
+	fi
 }
 
 # $1 jail name